feat: Added scrollbar
